feat(test-engine): [STABLE] 1B+1C+1E: integer division error tests + toThrow(Class, code) (ref #200)
1E: expect(fn).toThrow(ErrorClass, codeOrSubstring?) — now checks error
message contains the diagnostic code. Error bag handling improved in run().

1B+1C: integer-errors.test.ts — 40 new tests:
- E401 (division by zero): all 8 integer types × / and %
- E402 (MIN/-1 overflow): all 4 signed types × / and %
- Control values: all 8 types × / and % (7/2=3, 7%2=1)

Bug fix: E402 guards extended to i8/i16 (minMap was missing INT8_MIN,
INT16_MIN). Division temp variable type now matches operand type instead
of hardcoded int32_t (operators.ts + assign.ts, 4 locations).

1932 spec tests + 206 engine tests pass.

+++ packages/tests/test/cases/13-build/strict/safe-math/ok-u8-div-zero/expected.c
@@ -6,7 +6,7 @@ int main(void) {
     uint8_t b = 0U;
     uint8_t z = 0;
     MathError _math_err_0 = {0};
-    int32_t _math_1 = b;
+    uint8_t _math_1 = b;
     if (_math_1 == 0) { _math_err_0.operation = "div"; goto _catch_0; }
     z = a / _math_1;
     goto _catch_end_0;
